ProfileScreen [nfc]: Inline AwayStatusSwitch
